barbershop/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.7
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
regex==2023.6.3 